pyproject.toml
[project]
name = "searchess-ai-service"
version = "0.1.0"
description = "AI service for Searchess."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
]

[project.optional-dependencies]
# Install to use dataset pipeline, training, and supervised inference backend.
# torch is listed separately because it requires an index-url on some platforms.
training = [
    "python-chess>=1.999",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "torch>=2.0",
]

[project.scripts]
prepare-dataset = "searchess_ai.data_pipeline.pipeline:main"
train-model = "searchess_ai.training.pipeline:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pylint>=4.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/searchess_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
